+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-v2.4.7
+v2.4.7 (2023-03-04)
 ------
 * Fixed a bug in the override method which failed to look for super class interfaces.
 * Added methods to obtain the FieldMaker name, the MethodMaker name, and the MethodMaker